move method to parent
