+++ openpype/hosts/maya/api/plugin.py
@@ -5,6 +5,7 @@ from maya import cmds
 from avalon import api
 from avalon.vendor import qargparse
 from openpype.api import PypeCreatorMixin
+from avalon.pipeline import AVALON_CONTAINER_ID
 
 from .pipeline import containerise
 from . import lib
@@ -168,16 +169,18 @@ class ReferenceLoader(Loader):
                 return
 
             ref_node = get_reference_node(nodes, self.log)
-            loaded_containers.append(containerise(
+            container = containerise(
                 name=name,
                 namespace=namespace,
                 nodes=[ref_node],
                 context=context,
                 loader=self.__class__.__name__
-            ))
-
+            )
+            loaded_containers.append(container)
+            self._organize_containers([ref_node], container)
             c += 1
             namespace = None
+
         return loaded_containers
 
     def process_reference(self, context, name, namespace, data):
@@ -310,3 +313,13 @@ class ReferenceLoader(Loader):
                            deleteNamespaceContent=True)
         except RuntimeError:
             pass
+
+    @staticmethod
+    def _organize_containers(nodes, container):
+        # type: (list, str) -> None
+        for node in nodes:
+            id_attr = "{}.id".format(node)
+            if not cmds.attributeQuery("id", node=node, exists=True):
+                continue
+            if cmds.getAttr(id_attr) == AVALON_CONTAINER_ID:
+                cmds.sets(node, forceElement=container)
